Finished Report/Added NSEW Smart Integration
Availabe in PDF and docx. Added in NS priority mode and NS smart mode
that interface with the NSEW sensors. The FPGA bitfile needs to be
recompiled and the main_RT.vi needs to be updated for the additional
inputs. They are already layed out and only need to be connected after
the bitfile is updated.